Better error handling for tasks.
HPC-GAP now supports suppression of error message output (via the
thread-local variable SilentErrors, to be used in conjunction with
BreakOnErrors) and recovery of the last error message in that case with
the thread-local variable LastErrorMessage.

This is used by the standard task library to learn whether the task
succeeded or failed and why (via TaskSuccess() and TaskError()).
Likewise, task errors will now not drop the user into a break loop
anymore, though we may want to make this a configurable option.