Add NormalHallSubgroups
This is the second installment in the series of enhancing
StructureDescription. (The first being DirectFactorsOfGroup.)

A new global function NormalHalSubgroupFromSylows is introduced, which
computes normal Hall subgroups by computing normal closures of all Sylows.
An optional second argument "any" is recognized, in which case the function
returns the first nontrivial normal Hall subgroup, if exists, and returns
fail, otherwise. This will be used for computing semidirect decompositions,
because in this situation the Schur-Zassenhaus theorem guarantees a
complement.

NormalHallSubgroups is a new attribute, which consists of a list of all
normal Hall subgroups of a group. This is currently defined only for groups
which can compute size for all their subgroups. This check might not be
needed, e.g. the function works for DihedralGroup(IsFpGroup, 60), as well.

A test file is added.