Add ComplementClassesRepresentatives for all groups
This is the third installment in the series of enhancing
StructureDescription. (The first being DirectFactorsOfGroup, the second is
NormalHallSubgroups.)

This commit adds a method to ComplementClassesRepresentatives by
computing all conjugacy classes of subgroups, and then checking them
one-by-one for a suitable complement. It has a very low rank, and (just in
case) checks first whether some faster method cannot be used. Further, it
warns the user that all subgroups are computed, because this method can be
really inefficient.

In the fourth installment, SemidirectDecompositions will be written to use
ComplementClassesRepresentatives to determine whether or not a group can
be factored as a semidirect product.

Moreover, added a new KeyDependentOperation named
ComplementClassesRepresentativeKD, which calls
ComplementClassesRepresentative, but stores the result as a usual
KeyDependentOperation does. I believe this can be useful, and it should go
to the manual, but I would need help with that.

Added test file, which at least runs through the new code, and parts of
the old code, as well.