Rewrite new MaximalNormalSubgroups to work for infinite groups
New MaximalNormalSubgroups is rewritten to work for infinite groups, as well.
If G/G' is infinite, then there are infinitely many maximal normal subgroups,
and the method errors out. If G is nonabelian, then maximal normal subgroups
of G/G' are pulled back. If G is Abelian, then first it is represented as a
pc-group, then NormalMaximalSubgroups are computed, then pulled back to the
original group.

The tst file is also rewritten to test some infinite groups.