Rewrite DerivedSeriesOfGroup to check for abelian derived subgroups
DerivedSeriesOfGroup is rewritten to check for the derived subgroup being
abelian. In some cases this may terminate for fp-groups for which the old
method did not, in the expense of an extra IsAbelian check for each term.
Insted of while do, repeat until is used to decrease code repetition.

Test file is added.