Add MinimalNormalSubgroups and Socle for all nilpotent groups
For (finite or infinite) nilpotent groups every normal subgroup intersects
the center nontrivially, thus minimal normal subgroups are central. As
every central subgroup is normal, minimal normal subgroups are cyclic of
prime order. Hence every minimal normal subgroup is in fact a minimal normal
subgroup of the socle, where socle is the Omega group of the Center for all
prime divisors.

Thus Socle method for finite nilpotent groups is rewritten for all nilpotent
groups (test file contains examples of pcp and fp groups) using
AbelianInvariants and IndependentGeneratorsOfGroup. Rank is increased
so that this method would trigger before the finite solvable method.
Trivial methods for elementary abelian and simple groups are added.

New method for MinimalNormalSubgroups goes over the Sylows of the Socle and
makes note of all minimal subgroups. Rank is increased to run this method
first.

Tests are added.