Remove NameIsomorphismClass
Remove the undocumented attribute "NameIsomorphismClass". It was added
2000-01-05 by Hans Ulrich Besche, and always was marked as experimental.
Furthermore, it was only available as a pre-computed value for pc groups
constructed via `CyclicGroup`, and some groups constructed via `SmallGroup`.

As a replacement, you can use StructureDescription.

The only user visible change is that the output of `SmallGroupsInformation`
for products of three distinct primes does not list the isomorphism classes
anymore. But since other `SmallGroupsInformation` methods also don't do
that, this seems like an acceptable loss.