Clean up use of IsGeneratorsOfSemigroup
IsGeneratorsOfSemigroup was used inappropriately in invsgp.g* and in
semigrp.g*.

This commit includes cleaned up versions of the functions Semigroup,
Monoid, InverseSemigroup, and InverseMonoid. Some unnecessary code was
removed, and some changes, described below, were introduced.

In invsgp.g* it was used as a filter for the methods
InverseSemigroupByGenerators and InverseMonoidByGenerators. But since
collections are often not attribute storing, it was not possible to
enter any of these methods except with a collection in
IsAssociativeElementCollection. Note that since matrices are not in
IsAssociativeElementCollection, there are instances when the methods
should have applied but did not.

In semigrp.g*, IsGeneratorsOfSemigroup was checked for every one of the
arguments to the function Semigroup, and Monoid, but not to the final
list of generators obtained. For example, a collection of matrices of
different dimensions each individually satisfy IsGeneratorsOfSemigroup
but collectively do not. This commit introduces checks for the final
list of generators inside the functions Semigroup and Monoid.