Implement hex escapes, refactor GetStr and GetChar
This commit allows hexadecimal escapes of the form `0xHH` in string
and character literals where `H` is a hexadecimal digit.

The common codepaths for parsing the escape sequences in GetStr and GetChar
are factored out into a common function GetEscapedChar.

If an invalid escape sequence is read, a SyntaxWarning is displayed, but
the old and documented behaviour to just ignore the backslash is preserved.