Force a finiteness test in generic `PcgsElementaryAbelianSeries`
A test for the crisp package assumes that PcgsElementaryAbelianSeries can be
called for an fp group that does not yet know to be finite, and the
solvability test does not imply finite. As the main library cannot edit this
package test, this needs to be changed so the standard test suite runs
through.