Added 'testpackage' target to Makefile
Usage:
make testpackage PKGNAME=<pkgname>

This will run standard tests of the package <pkgname> in the same
way like we run them in our regression testing (though with the
user's selection of packages that may be different). It will write
test output to 3 log files dev/log/testpackageX_timestamp.pkgname
where X=1 with no packages, 2 with all packages and A with default
packages.