buildsys: handle GAP version
The approach we use is based on what git.git does.
It stores the generated version in a separate file cnf/GAP-VERSION-FILE,
which makes it possible to cleverly handle rebuilding files
that use the version. It also makes it possible to use version
in non-C-code later one, should we need to.