Fix AsPartialPerm method for transformations
The previous method was written when transformations had a fixed degree,
and was incorrect when applied to a degree-less transformation (i.e. all
transformations in GAP since GAP 4.7).

The method, related tests and doc, for AsPartialPerm of a transformation
(and no further args) were removed. It is not very clear what such a
method should do.