Fix nested timeouts
If the remaining timeout in a nested call was too close, i.e. less than
a microsecond away, a division of nanoseconds by 1000 would yield a timeout
of zero which would lead to the timeout code getting stuck in a loop. Clamp
the reset value to microsecond to work around this.