Fix #397
If the interpreter runs into an error while reading from a stream and
tries to check for a dual semicolon at the end of the statement, the
call to GetSymbol() leads to a call to ReadLine on the input stream.

With  (IsInputTextStream and IsInputTextStringRep) the correct method
is selected and called. The crash happens inside the method when trying
to access the stream or local variables.

This might not be the correct fix yet, but it seems clear that the
interpreter state is not right at this point.