INSTALL.md: don't refer to "Unix and OS X", don't say "UNIX"
The term UNIX is trademarked and refers to *the* UNIX operating system,
but not its derivates.

Mac OS X actually is a direct descendant of a version UNIX, and thus
directly related to UNIX (it is derived from FreeBSD, which in turn is
derived from BSD UNIX). Thus it seems weird to refer to "Unix and OS X".
However, some OS X users may not realize that a section about "Unix"
is relevant for them, so I sometimes used expressions like "Unix
(including OS X)".

By the way, note that Linux (which most people will consider a Unix)
strictly speaking is NOT a direct member of the Unix family: While
developed to be compatible with UNIX, it (at least originally) shares
no code with it, and thus is a "UNIX clone". So in a sense, it is a
less "a Unix" than OS X. Of course that just shows how silly that
distinction is, nothing else. I just mention it to reinforce that if
we call Linux a Unix, then we should do so the same for OS X.