Alternative to handle functions returning no value
This is an experiment to allow users statements like
  a := fu();
even if fu() does not return a value. In this case the statement
unbinds 'a' so that one can detect that fu() did not return a value
by 'IsBound(a);'.

The behaviour is similar for assignments to lists or record components
where statements like
  l[i] := fu();
  r.comp := fu();
in case of fu() not returning a value have the effect of
  Unbind(l[i]);
  Unbind(r.comp);