Make IsomorphismFpMonoid an attribute of semigroups
A semigroup can be isomorphic to a monoid, while not being a monoid in
the technical GAP sense. For example, the semigroup generated by
Transformation([1, 2, 3, 3, 3]) is not a monoid in the GAP sense but is
mathematically.  With this change we could now install methods for
IsomorphismFpMonoid for such a semigroup.