Fix One/Zero/MultiplicativeZero for partial perms
There were several issues here:

1. some declarations were missing

2. for a partial perm semigroup without generators the default method or
a One of a partial perm collection was used. This always returned an
answer, when it shouldn't have it if the One was not in the semigroup.

3. partial perms and semigroups of partial perms had methods for Zero,
which according to the manual "return the additive neutral element of
the additive element". Since partial perms are not additive these
methods were inappropriate. As such they are removed in this PR and
methods for MultiplicativeZero are introduced instead.

This commit fixes these issues.