Fixed issue 820 and added tests
Fixed List so that it complains if it has more than two arguments. Also changed the error type to 'ErrorNoReturn' when the arguments are 0 or greater than 2. Added some additional tests to test the error messages in these cases, and to test some of the other documented features of List, that is, it works for lists with holes, and it returns a mutable list.