ENHANCE: In PcSeries do not induce if the pcgs is being built
There are two different situations in which PcSeries is called. The first is
a good pcgs and we happen to want the series -- this is the existing code.

The second case is that the series is needed to do anything with the pcgs.
In this case inducing caused bad recursions. Avoid this.