FIX: ClosurePermGroup long run
In some situations, it seems that ClosurePermGroup can iterate very long
time with evrification failing repeatedly. In this situation we now simply
force a new stabilizer chain computation from scratch, than trying to bend
the (apparently confused) chain in the right way.
This is not a beautiful fix but a workaround, albeit one that should have
litlle cost implications.