Handle functions returning no value correctly
This patch does not fix the GAP -> C compiler.
I suspect there could be many "modern" GAP features that are broken in
the GAP -> C compiler.

Technical explanation:

In 4.8 and before, GAP would decide if it would need to the return
value of a function straight after the function call was finished. This
worked fine then, because it was easy to deduce if we would need the
return value. It did fail in some places, for example GAP 4.8 would not
allow (f()) if f() did not return a value. With a relaxing of the GAP
grammar in gap-master, this has gotten much harder to do.

Now we switch to a new method -- instead track when functions have
returned nothing, and then later if we try to "use" the return value of
that function, then throw an error. This involves putting a special
value on the stack to mark functions which returned nothing (which were
previously indistiguisable from any of the many other operations which
return nothing), and producing an error if the value is ever used. This
magic value is never seen outside a very small set of functions.

This does change GAP's behaviour very slightly. For example as mentioned
above (f()) in 4.8 failed if f did not return a value. Now it works fine.

The main difference is that GAP now delays noticing when a function
didn't return. For example, given:

f := function() Print("X"); end;
Maximum(f(),f(),f());
GAP 4.8 prints X, with this patch master prints XXX, as we invoke all
the calls to f before we then try and call Maximum, which notices the
empty values. I don't expect this to cause a serious issue, and the
effects are quite limited (and only effects code which is going to
throw an error anyway).