FIX: Maximal subgroups for S_n/A_n pingpong between prim/tom datalibs.
Clean up the way how the `maximal subgroups from table of marks' can refer
to the `maximal subgroups of S_n/A_n by primitive groups.
Move that maximal subgroups code from prim into the main library (so it is
in one place and its use is clear.

This fixes #984