Log every package to different file with Buildpackages.sh
Packages failed to build are logged to pkg/log/fail.log, together with a
date of when they failed.
Every package PKG build output is logged to a file in pkg/log/$PKG.log.
The whole output is written into pkg/log/buildpackages.log.
If there is stderr output, it is written to .err files, then the stdout
output is written to .out files. The .log files contain the whole output
in sequence.
Every .log, .out, .err file is overwritten at every rerun of
buildpackages.sh, ***except*** for fail.log, which is always appended
together with the date of the run.
Output of ls is dangerous to parse, replaced by find.