zmodnz: always use ZmodnZObj to create instances
Instead of sprinkling calls to Objectify throughout the code, funnel
all creation of ZmodnZObj objects through the constructor functions
named ZmodnZObj(). In a very few cases, this means we computed "mod"
twice (namely for quotients, and powers), but the overhead for
doing this to an already reduced element is negligible.

The advantage of this is of course reduced code duplication, and it now
becomes easier to make changes to the creation of instances in a uniform
way, e.g. we could now easily make all ZmodnZObj instances immutable.