meataxe: improve SMTX.DualizedBasis
The old code created five matrices: the transpose, its nullspace (which
internally makes a copy of its input), a copy of the nullspace,
and finally an immutable copy of the triangulized nullspace.

The new code creates just three: first a mutable transpose,
and then we invoke TriangulizedNullspaceMatDestructive on that, which
can work in place, and returns the triangulized nullspace,
of which we make an immutable copy.