Enhance HasXXFactorGroup
HasSolvableFactorGroup computes commutators iteratively, and checks if
they are in the normal subgroup. If DerivedSeries is already known,
then checks if its last term is in N.
HasAbelianFactorGroup and HasElementaryAbelianFactorGroup returns true if
group is already known to be abelian or elementary abelian. The latter
previously errored out if normal subgroup was the whole group, fixed.
Tests are added.