gmpints: pedantic tweak to ModInt
This change should be mostly irrelevant.; indeed, the C99
standard mandates that a%b and a%(-b) are equal, and even
on most pre-C99 compiler this holds true -- but not all.

We could instead also add a configure test, and reject compilers
violating this; in that case, we can obviously just stop checking
the sign of k.