gmpints: add POWERMODINT and INVMODINT
This adds a GMP based implementation of PowerModInt. The new kernel
function INVMODINT computes modular inverse, and is not currently
exposed to the library in a documented function.

As usual, we get a speedup. Old code:

gap> for b in [-100..100] do for e in [1..100] do for m in [1..100] do
>  PowerModInt(b,e,m);od;od;od; time;
2270
gap> for b in [-100..100] do for e in [1..100] do for m in [1..100] do
>  PowerModInt(b,2^e,m);od;od;od; time;
11354

New code:

gap> for b in [-100..100] do for e in [1..100] do for m in [1..100] do
>  PowerModInt(b,e,m);od;od;od; time;
625
gap> for b in [-100..100] do for e in [1..100] do for m in [1..100] do
>  PowerModInt(b,2^e,m);od;od;od; time;
1920