gmpints: fix RemInt error handling
Surely the user does not want to retry as QUO / QuoInt...

Note: Just changing QUO to RemInt and the `if` to `while` would not be
enough; we'd also have to verify that opR stays an integer. And then
adjust the error message. And so on...

Since I am doubtful regarding the usefulness of ErrorReturnObj to start
with, I chose to keep my life simple and use ErrorMayQuit instead.