gmpints: add JACOBI_INT and change Jacobi() to use it
Use GMP's mpz_kronecker to compute the Kronecker-Jacobi symbol J(n,m).
We retain the old GAP implementation for reference, and also to
facilitate testing (we want to make sure that old and new code return
identical results).

Note that the new implementation accepts some more inputs, namely it
allows negative values for m. This is because GMP allows this, and
it seems pointless to restrict this again.

The new implementation is considerably faster. Old code:

gap> for a in [1..1000] do for b in [1..1000] do Jacob(a,b);od;od;time;
1979

New code:

gap> for a in [1..1000] do for b in [1..1000] do Jacob(a,b);od;od;time;
85