kernel: fix bug in LtTrans24 in trans.c
This commit fixes a bug in LtTrans24, a for loop was incorrectly nested.
In order to increase the test coverage to include this part of the code
(and a number of other parts of the code in trans.c that were previously
unreachable) IdentityTrans is altered to be a T_TRANS4. Tests are added
for this part of the code, and the other previously unreachable parts.