kernel: removed more HAVE_FOO_H macros
In particular, removed more of these were we either #include the file
elsewhere without a guard anyway; or where we have no alternative
implementation that works without the header.

We should only add these back if / when we encounter platforms where
they make a difference; and in that case, they should be added back with
a comment pointing out the systems that need them!