Do not reset tilde in ClearError
This is undoing an earlier fix. The problem was that the ~ variable is not
cleared when exiting a break loop, or after an error. Reseting it's value
in ClearError fixed this, but it turns out ClearError is called even when there
are no errors, for example after reading a file.

This does reintroduce a tilde problem -- but it will only be hit by programs
which cause an error, then try to read tilde and find it has a value,
when it shouldn't.