tests: rename files in testbugfix, and split a few into two
The new names put the date of the bugfix into the filename.
This is mainly useful for future bugfixes, as it helps to avoid
clashes between multiple pull requests adding new bugfix tests.

For future bugfix .tst files, I suggest we use the format
  YYYY-MM-DD-short_description.tst
for the filename. For now, the old tests use the old test number
as "description".