Speed up Root(s)Mod for moduli with large prime factors
For RootMod and RootsMod, all computations are eventually reduced to
prime moduli. This means we have to factor the initial modulus, which
can be very expensive. However, it actually suffices if the factors are
Fermat pseudo primes relative to certain bases... So we switch to using
the UseProbabilisticPrimalityTest option, and add a test which verifies
the required property n^(p-1) = 1 (mod p) where necessary. If this
fails, we could do something clever to recover, but instead, we simply
print an error message requesting that the user reports it to use; for
if it happens, then either there is a bug (which we should fix), or else
we found a BPSW pseudo prime that is not a prime, and this is something
we definitely would want to know about (and publish, too ;-).

Fixes #758