kernel: introduce GAP_PATH_MAX
Some places in our code were limited to paths of length 256, some had
512 as limit. Now they all use the new constant GAP_PATH_MAX as limit,
which is set to 512 for now, but could be increased later on.

Note that most operating systems don't actually impose a limit on path
lengths (and if they do, it tends to be much larger), so this isn't the
final word. But it's at least a start.

We also extend SyFindGapRootFile to support output buffers of arbitrary
size, and adapt SyFindOrLinkGapRootFile accordingly. Also cleanup
SyFindOrLinkGapRootFile and remove various redundant buffers and copying
operations.