Add config.guess, config.sub and install-sh to repository
We are not using automake, but autoconf wants to use config.guess, config.sub
and install-sh, which are part of automake. This leads to an error in
autogen.sh (which one might classify as a bug in autoreconf).

As a workaround, we add our own copies of these files (as an added benefit,
they are newer than what is currently shipped with automake).

In the future, we may wish to revert this, and explicitly instruct autotools
to copy the files for us again.