hpcgap/tst: remove multi-arg lambda tests
Support for these has not yet been ported to HPC-GAP. To remove this
test for HPC-GAP, we add a new test file
  hpcgap/tst/testinstall/function.tst
which started out as a copy of
  tst/testinstall/function.tst
but with the relevant tests removed. Once multi-arg lambdas are
supported in HPC-GAP, the new function.tst file can be removed again.