TestDirectory: accept list of list of directories as argument
Given input
  [  [dirA1, dirA2, ...], [dirB1, dirB2, ...], ... ]
we treat [dirA1, dirA2, ...] similar to how Filename() does: We first gather the union
of names of *.tst files in those directories. Then, for each filename, we iterate over
dirA1, dirA2, ... until we find a directory that actually contains a file with that
name, and add it to the list of test files. Then we go to the next filename.

Then we move to [dirB1, dirB2, ...] and repeat the process.

This is meant to allow support for testing files across multiple GAP roots
(as e.g. happens when HPC-GAP is enables): If both hpcgap/tst/testinstall/FOO.tst
and tst/testinstall/FOO.tst are present, we use the former. If only the latter
is present, we use that. This way, if both files exist but are identical, we can
remove the copy in hpcgap/tst without affecting the tests.