src/system.c: remove SyTime implementation using times()
First off, it was incorrect: it used CLOCKS_PER_SEC instead of
sysconf(_SC_CLK_TCK), which is my fault (made in 2012). Luckily,
on Cygwin, this seems to make no difference, and that is the only
platform we still used this code.

The reasons we used times() in cygwin is that getrusage does not track
child processes. But as it turns out, neither does times()! So there
is no reason to keep that code.