Retry "make bootstrap" on CI servers a few times
Travis builds sometimes fail on the `make bootstrap-pkg-full` step, due
to an "Connection refused" error from the server. This is likely caused
by the server being overloaded (a dozen or so CI builds start
simultaneously and if things align badly, will try to perform the
boostrap around the same time.

With this change, we instruct wget to retry the download up to 5 times,
with 5 seconds between each try.

As a side effect, user's can in general override the arguments we pass
to wget (and even switch things to use e.g. curl instead).