src/objects.h: add NEXT_EVEN_INT and use it to define TNUMs
Specifically, certain TNUMs must always be even. Thus when an odd number
of TNUMs is inserted before them, they need to be adjusted, which so far
required manual changes, which was easy to forget.

With NEXT_EVEN_INT, this is now automatic; moreover, it is now evident
that these TNUMs must be even.

In addition, FIRST_COPYING_TNUM now also uses NEXT_EVEN_INT, which it
previously didn't. This ensures that mutability checks still work even
if for objects being copied.

Finally, in the definition of COPYING, we replace FIRST_RECORD_TNUM by
FIRST_IMM_MUT_TNUM; while the two are equal in value, the latter conveys
the appropriate semantic information in this use case.