kernel: fix early ErrorInner version
On 2008-01-18, ErrorInner was changed from a function taking six args, to a
function taking two args (a record and a string). However, our early
ErrorInner implementation was not adapted to that.

Ironically, this was already fixed on the HPC-GAP side, but I accidentally
removed the fix there in an earlier HPCGAP/GAP sync, as I did not realize
its meaning back then.