All help matches pointing to same subsection
Sometimes the help system finds several matches for a query, but
all of them actually point to the same subsection of the same book
(e.g., when several functions are documented in one ManSection
of a GAPDoc manual). With this patch the corresponding subsection
will be displayed directly (instead of offering the user a choice).

To implement this a new utility function .SubsectionNumber is
needed from the help book handlers.

For the "default" handler (old gapmacro.tex manuals) this new function
is also contained here. For GAPDoc manuals this will be provided with
the upcoming GAPDoc 1.6 release.

There is a sensible fallback if a help book handler does not provide
the new utility function.