Add a basic implementation of thread-local attributes, for HPC-GAP
Thread-local attributes provide a way for an object to have
a seperate copy of mutable attributes, such as StabChainMutable,
in each thread.

The implementation is designed to be simple, rather than provide a
complete implementation of all features of attributes, in particular
thread-local attributes can not be used as filters of methods or
operations.