kernel: add check to ON_KERNEL_ANTI_ACTION
This commit adds a check that the argument ker has zero length, and if
it does, just returns it. Previously if ker was of length 0, then the
behaviour was undefined since the value of ELM_LIST(ker, 1) was used but
not defined. Also added an assert to the start of the function.