kernel: rename MARK_BAG to MarkBag and turn it into a function
The function is *not* static inline, so that we can hide more
implementation details inside gasman.c, instead of exposing them to the
rest of the kernel. This has a hypothetical tiny performance impact
on marking functions; however, for the default marking functions we
avoid the problem by providing an internal `static inline` version of
MarkBag called MarkBagIntern; and it turns out that not much code
uses non-standard marking functions.

The next commit will replace a few more uses of MARK_BAG with optimized
internal marking functions.