kernel: introduce MarkArrayOfBags convenience function
This allows marking a bunch of bags at once. It is slightly more
efficient than directly iterating over the array and invoking MarkBag on
each entry, as it is in the same translation unit as MarkBag, and thus
the compiler can inline the MarkBag calls.

Also made MarkAllSubBags and MarkAllSubBagsDefault use it to reduce code
duplication.