buildsys: remove code dealing with HPC-GAP headers
Now that there are no header files left in hpcgap/src/, we can remove all code
that dealt with them.

This also enhances compatibility with packages that build kernel extensions.
In particular, we can now build edim for HPC-GAP again under Travis. (which
doesn't mean it works, but at least it is one step closer to working)