MatrixObj: rename IsRowVectorObj -> IsVectorObj
We decided that vector objects should not distinguish between row or column
orientation. If you need a "column vector" or "row vector", then use a n*1 or
1*n matrix object instead.

We keep IsRowVectorObj as a synonym around, in order to ensure that code
using it keeps working (most notably, the cvec package). However, we may
remove this in the future.