kernel: merge Elm2List/ELM2_LIST, Ass2List/ASS2_LIST, change behavior
While we still want a hotpath, this is not the way we want to do it.

Also, we now dispatch ELM2_LIST to the operation \[\] (= ELM_LIST)
with three arguments (list,pos1,pos2) instead of two (list,[pos1,pos2]),
and similar for ASS2_LIST.

This makes method installations easier to read, and avoids the extra
overhead caused by creating a plist just to wrap the two arguments.