lib/list.gd: restrict [], []:=, IsBound, Unbind to pos int indices
This is how it always was before we made it more generous in 2015 to
allow for fancy index operations (like `matrix[1,2]`, or `dict["key"]`).
We still support that, but code which wants to do it should either
use `InstallOtherMethod` (like the tests), or use `DeclareOperation` to
explicitly declare the supported index types.

This way, incorrectly using `InstallMethod` by accident at least triggers
a GAP warning / error.