Enable -O2 on travis
GAP is by default built using -O2, but on Travis, we set custom CFLAGS, which
unintentionally overrode this default. This PR turns -O2 back on, which
results in tests completing up to two times faster.

However, this leads to vecgf2.c triggering an internal compiler error in
the 32bit HPC-GAP build. So we turn off -O2 for that.