Fix two issues in FuncSTRONGLY_CONNECTED_COMPONENTS_DIGRAPH
SET_PLIST_ELM is equivalent to setting ((UInt *)ADDR_OBJ(val))[x], but
we should not be using PLIST functions on a DATOBJ, as the new kernel
debugging rejects it.

Also this line should not use INTOBJ_INT, as 'val' contains normal
C ints. However, all this particular int is used for it to denote a value
larger than n, and INTOBJ_INT(n+1) is 4*(n+1)+1, so the code still
worked fine.