travis: reorder jobs, and drop redundant testinstall jobs
The new order tries to put generally slower jobs to the start, and
faster ones near the end. The hope is that this will decrease the
wall clock time used to complete a full Travis build, as we avoid
the currently typical situation were 1-2 rather long running jobs
are only started after almost all other jobs have completed, thus
making suboptimal use of the multiple build jobs Travis allows us
to run in parallel.

In addition, we avoid building custom GMP in most cases (as it is
very slow).