hpcgap: remove FIXED_REGION object flag
This reverts commit 941d2d252c34bdf27e2667d5de20a0fa44a8425d. Fixed
regions are not currently used by anything; they were / are intended for
interacting with external parallelized code.

In any case, the current implementation is certainly incomplete; e.g.
all kinds of actions currently reset object flags, which violates the
purpose of the FIXED_REGION flag.

If / when this feature is actually needed in the future, we can of
course reintroduce it.