Remove CLI option -X and obsolete CRC checking code
This also removes the global variable SyCheckCRCCompiledModule.
Note: This does not remove the CRC checks for GAP files compiled to C
cod by GAC. Rather, the removed code was a left-over from the long-gone
"completion" mechanism. It did not actually do anything, because the
only call to SyFindOrLinkGapRootFile hardcoded 0 as reference CRC.