Tweak tst/testinstall/hpc/threads.tst to avoid cygwin bug
On cygwin, usleep(), which is supposed to sleep the process for a given number
of *micro*seconds, ends up sleeping the process several *milli*seconds longer.
As a result, this test took 25+ minutes instead of less than a second.

We avoid this by using a larger sleep duration but fewer iterations.