kernel: change `a[i,j,k]` with 3+ indices to syntax error
For now, we only support 1 or 2 indices. If we want to, we can add back
support for 3+ indices in the future, with the caveat that we'll either
need to pass them in a plist (which makes this case differ from the 1
and 2 indices case); or else we are limited to at most 4 indices,
as method dispatch for operations supports at most 6 args right now,
and we need one for the "matrix", and another for the an element
to insert into the matrix at position described by the indices.

Or we could implement support for operations with more than 6 args...