gvars: fix a potential race condition in GVarName
Namely, we compute the hash of the variable name modulo SizeGVars.
If the lookup fails, we briefly release the gvars lock, in order to
re-acquire it as a full read-write lock. If during this time, the
value of SizeGVars changes, we may end up placing the gvar in the
wrong position.

As a byproduct, this fixes a compilation warning (resolves #1459).