records: fix a potential race condition in RNamName
Namely, we compute the hash of some string modulo SizeRNam. But we did
so before calling HPC_LockNames(); thus a concurrent thread may lead
to SizeRNam being modified before we take the lock.

Moreover, if the lookup fails, we briefly release the lock, in order
to re-acquire it as a full read-write lock, which again opens us up to
a race condition.

Also fix a typo and change some tabs to spaces