Avoid irreducibility test in `AlgebraidExtension`
AlgebraicExtension is mainly an administrative function. It should not
spend considerable time with testing its input polynomial for irreducibility
(it is an error anyway if the polynomial is reducible, this should be tested
by the user beforehand, or be clear from the mathematical context).

This pull request substitutes the irreducibility test by the much cheaper
test, if the given polynomial has coefficients in the base field.
This test seems sensible, for example in case of iterated algebraic
extensions, where a user may want to use a polynomial over a subfield which
first needs to be rewritten over the larger field.