Use #ifdef, not #if, to check for autoconf HAVE_* macros
Technically, both work, as autoconf always does a

  #define HAVE_FOO 1

However, autoconf indicates absence of a feature by *not* defining
HAVE_FOO (as opposed to defining it to be 0). Thus, doing `#if HAVE_FOO`
triggers a warning if -Wundef is used. In order to distinguish valid
checks from those accidentally testing for an undefined macro, check via
`#ifdef HAVE_FOO` instead.